obsolete_kuterless_archieves/linux__current_requirement.txt
Django==1.6.1
Fabric==1.8.1
Flask==0.10.1
Jinja2==2.7.1
Markdown==2.3.1
MarkupSafe==0.18
Pillow==2.2.2
South==0.8.4
Werkzeug==0.9.4
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-allauth==0.17.0
django-classy-tags==0.4
django-cors-headers==1.1.0
django-extensions==1.2.5
django-filter==0.7
django-floppyforms==1.1.1
django-model-utils==2.0.3
django-mptt==0.6.0
django-notifications-hq==0.6.2
django-password-reset==0.7
django-sekizai==0.7
django-taggit==0.12.3
django-toolbelt==0.0.1
djangorestframework==2.3.14
ecdsa==0.10
gunicorn==18.0
ipython==1.1.0
itsdangerous==0.23
oauthlib==0.6.3
paramiko==1.10.0
psycopg2==2.5.1
pycrypto==2.6.1
pyreadline==2.0
python-openid==2.2.5
pytz==2013.8
requests==2.3.0
requests-oauthlib==0.4.1
six==1.4.1
sorl-thumbnail==11.12
static==0.4
wiki==0.0.23
wsgiref==0.1.2
